requirements.txt
pandas==2.2.3
mysql-connector-python==9.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
